Improve scripts to only install things that haven't been installed before

+++ build_scripts/ubuntu_sim.sh
@@ -4,20 +4,31 @@
 sudo apt-get remove modemmanager -y
 
 # Ninja build system
-mkdir -p $HOME/ninja
-cd $HOME/ninja
-wget https://github.com/martine/ninja/releases/download/v1.6.0/ninja-linux.zip
-unzip ninja-linux.zip
-rm ninja-linux.zip
-exportline="export PATH=$HOME/ninja:\$PATH"
-if grep -Fxq "$exportline" ~/.profile; then echo nothing to do ; else echo $exportline >> ~/.profile; fi
-. ~/.profile
+ninja_dir=$HOME/ninja
+echo "Installing Ninja to: $ninja_dir."
+if [ -d "$ninja_dir" ]
+then
+    echo " Ninja already installed."
+else
+    pushd .
+    mkdir -p $ninja_dir
+    cd $ninja_dir
+    wget https://github.com/martine/ninja/releases/download/v1.6.0/ninja-linux.zip
+    unzip ninja-linux.zip
+    rm ninja-linux.zip
+    exportline="export PATH=$ninja_dir:\$PATH"
+    if grep -Fxq "$exportline" ~/.profile; then echo " Ninja already in path" ; else echo $exportline >> ~/.profile; fi
+    . ~/.profile
+    popd
+fi
 
 # Common Dependencies
+echo "Installing common dependencies"
 sudo add-apt-repository ppa:george-edison55/cmake-3.x -y
 sudo apt-get update -y
 sudo apt-get install python-argparse git-core wget zip python-empy qtcreator cmake build-essential genromfs -y
 # required python packages
+sudo apt-get install python-dev -y
 sudo apt-get install python-pip
 sudo -H pip install pandas jinja2
 
@@ -25,6 +36,7 @@ sudo -H pip install pandas jinja2
 sudo apt-get install ant openjdk-8-jdk openjdk-8-jre -y
 
 # Gazebo simulator
+echo "Installing Gazebo8"
 sudo apt-get install protobuf-compiler libeigen3-dev libopencv-dev -y
 sudo sh -c 'echo "deb http://packages.osrfoundation.org/gazebo/ubuntu-stable `lsb_release -cs` main" > /etc/apt/sources.list.d/gazebo-stable.list'
 ## Setup keys
@@ -36,8 +48,17 @@ sudo apt-get install gazebo8 -y
 ## For developers (who work on top of Gazebo) one extra package
 sudo apt-get install libgazebo8-dev
 
+
 # Clone PX4/Firmware
-mkdir -p ~/src
-cd ~/src
-git clone https://github.com/PX4/Firmware.git
-cd Firmware
+clone_dir=~/src
+echo "Cloning PX4 to: $ninja_dir."
+if [ -d "$clone_dir" ]
+then
+    echo " Firmware already cloned."
+else
+    mkdir -p $clone_dir
+    cd ~/$clone_dir
+    git clone https://github.com/PX4/Firmware.git
+    cd Firmware
+fi
+cd $clone_dir/Firmware
